chore: finalize and clean up sponsor resume bank/dashboard
