+++ src/app/(auth)/(sponsor)/review/table/testData.ts
@@ -197,24 +197,6 @@ const testData = [
         '6': 'https://www.linkedin.com/in/johndoe',
         '7': 'https://x7pvefn5lz1gfty3.public.blob.vercel-storage.com/submissions/team-15/documentation-5cHPbbBU9fsXlvonAMnziOgw127juw.pdf',
     },
-    {
-        '1': 'Jane',
-        '2': 'Smith',
-        '3': 'University of British Columbia',
-        '4': 'Canada',
-        '5': 'https://www.github.com/janesmith',
-        '6': 'https://www.linkedin.com/in/janesmith',
-        '7': 'https://x7pvefn5lz1gfty3.public.blob.vercel-storage.com/submissions/team-15/documentation-5cHPbbBU9fsXlvonAMnziOgw127juw.pdf',
-    },
-    {
-        '1': 'Alice',
-        '2': 'Johnson',
-        '3': 'Harvard',
-        '4': 'USA',
-        '5': 'https://www.github.com/alicejohnson',
-        '6': 'https://www.linkedin.com/in/alicejohnson',
-        '7': 'https://x7pvefn5lz1gfty3.public.blob.vercel-storage.com/submissions/team-15/documentation-5cHPbbBU9fsXlvonAMnziOgw127juw.pdf',
-    },
 ];
 
 export default testData;
