+++ src/app/(auth)/(sponsor)/statistics/page.tsx
@@ -8,7 +8,7 @@ export default function StatisticsPage(props: NotFoundProps) {
     return (
         <div className="flex h-full w-full flex-col items-center justify-center gap-8">
             <Image
-                src="/login/sad-otter.webp"
+                src="/dashboard/messy-otters.webp"
                 width={699}
                 height={725}
                 className="max-w-[240px]"
@@ -19,11 +19,11 @@ export default function StatisticsPage(props: NotFoundProps) {
                     Gathering data...
                 </p>
                 <h1 className="mb-3 text-3xl leading-tight font-semibold text-balance text-white">
-                    This page will be available sometime after StormHacks 2025.
+                    This page will be available during StormHacks 2025.
                 </h1>
                 <p className="text-base text-balance text-white/60">
-                    Stormy and Sparky are currently working hard gathering the
-                    statistics!
+                    Stormy and Sparky are currently working hard gathering
+                    insights!
                 </p>
             </div>
             {props.children}
